[:integtest:pg] Refactor pg test
