+++ integtest/pg/src/test/java/io/github/zero88/integtest/jooqx/pg/jooq/PgJDBCRx2Test.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 
 import io.github.zero88.integtest.jooqx.pg.PgUseJooqType;
 import io.github.zero88.jooqx.DSLAdapter;
-import io.github.zero88.jooqx.rxjava3.Jooqx;
-import io.github.zero88.jooqx.rxjava3.JooqxBuilder;
+import io.github.zero88.jooqx.reactivex.Jooqx;
+import io.github.zero88.jooqx.reactivex.JooqxBuilder;
 import io.github.zero88.jooqx.spi.jdbc.JDBCErrorConverterProvider;
 import io.github.zero88.jooqx.spi.jdbc.JDBCPoolAgroalProvider;
 import io.github.zero88.jooqx.spi.pg.PgSQLJooqxTest;
@@ -17,20 +17,20 @@ import io.vertx.jdbcclient.JDBCPool;
 import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxTestContext;
 
-class PgReARx3Test extends PgSQLJooqxTest<JDBCPool>
+class PgJDBCRx2Test extends PgSQLJooqxTest<JDBCPool>
     implements PgUseJooqType, JDBCPoolAgroalProvider, JDBCErrorConverterProvider {
 
-    Jooqx jooqxRx3;
+    Jooqx jooqxRx2;
 
     @Override
     @BeforeEach
     public void tearUp(Vertx vertx, VertxTestContext ctx) {
         super.tearUp(vertx, ctx);
         this.prepareDatabase(ctx, this, connOpt, "pg_data/book_author.sql");
-        jooqxRx3 = JooqxBuilder.newInstance(io.github.zero88.jooqx.Jooqx.builder())
-                               .setVertx(io.vertx.rxjava3.core.Vertx.newInstance(vertx))
+        jooqxRx2 = JooqxBuilder.newInstance(io.github.zero88.jooqx.Jooqx.builder())
+                               .setVertx(io.vertx.reactivex.core.Vertx.newInstance(vertx))
                                .setDSL(jooqx.dsl())
-                               .setSqlClient(io.vertx.rxjava3.jdbcclient.JDBCPool.newInstance(jooqx.sqlClient()))
+                               .setSqlClient(io.vertx.reactivex.jdbcclient.JDBCPool.newInstance(jooqx.sqlClient()))
                                .build();
     }
 
@@ -38,7 +38,7 @@ class PgReARx3Test extends PgSQLJooqxTest<JDBCPool>
     void test_query(VertxTestContext ctx) {
         final Books table = schema().BOOKS;
         Checkpoint cp = ctx.checkpoint();
-        jooqxRx3.rxExecute(dsl -> dsl.selectFrom(table), DSLAdapter.fetchJsonRecords(table)).subscribe(recs -> {
+        jooqxRx2.rxExecute(dsl -> dsl.selectFrom(table), DSLAdapter.fetchJsonRecords(table)).subscribe(recs -> {
             ctx.verify(() -> Assertions.assertEquals(7, recs.size()));
             cp.flag();
         }, ctx::failNow);
