+++ integtest/pg/src/test/java/io/github/zero88/integtest/jooqx/pg/jooq/PgJDBCRx3Test.java
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.Test;
 
 import io.github.zero88.integtest.jooqx.pg.PgUseJooqType;
 import io.github.zero88.jooqx.DSLAdapter;
+import io.github.zero88.jooqx.rxjava3.Jooqx;
+import io.github.zero88.jooqx.rxjava3.JooqxBuilder;
 import io.github.zero88.jooqx.spi.jdbc.JDBCErrorConverterProvider;
 import io.github.zero88.jooqx.spi.jdbc.JDBCPoolAgroalProvider;
 import io.github.zero88.jooqx.spi.pg.PgSQLJooqxTest;
@@ -15,29 +17,28 @@ import io.vertx.jdbcclient.JDBCPool;
 import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxTestContext;
 
-class PgReARxMutinyTest extends PgSQLJooqxTest<JDBCPool>
+class PgJDBCRx3Test extends PgSQLJooqxTest<JDBCPool>
     implements PgUseJooqType, JDBCPoolAgroalProvider, JDBCErrorConverterProvider {
 
-    io.github.zero88.jooqx.mutiny.Jooqx jooqxMutiny;
+    Jooqx jooqxRx3;
 
     @Override
     @BeforeEach
     public void tearUp(Vertx vertx, VertxTestContext ctx) {
         super.tearUp(vertx, ctx);
         this.prepareDatabase(ctx, this, connOpt, "pg_data/book_author.sql");
-        io.vertx.mutiny.jdbcclient.JDBCPool pool = io.vertx.mutiny.jdbcclient.JDBCPool.newInstance(jooqx.sqlClient());
-        jooqxMutiny = io.github.zero88.jooqx.mutiny.JooqxBuilder.newInstance(io.github.zero88.jooqx.Jooqx.builder())
-                                                                .setVertx(io.vertx.mutiny.core.Vertx.newInstance(vertx))
-                                                                .setDSL(jooqx.dsl())
-                                                                .setSqlClient(pool)
-                                                                .build();
+        jooqxRx3 = JooqxBuilder.newInstance(io.github.zero88.jooqx.Jooqx.builder())
+                               .setVertx(io.vertx.rxjava3.core.Vertx.newInstance(vertx))
+                               .setDSL(jooqx.dsl())
+                               .setSqlClient(io.vertx.rxjava3.jdbcclient.JDBCPool.newInstance(jooqx.sqlClient()))
+                               .build();
     }
 
     @Test
     void test_query(VertxTestContext ctx) {
         final Books table = schema().BOOKS;
         Checkpoint cp = ctx.checkpoint();
-        jooqxMutiny.execute(dsl -> dsl.selectFrom(table), DSLAdapter.fetchJsonRecords(table)).subscribe().with(recs -> {
+        jooqxRx3.rxExecute(dsl -> dsl.selectFrom(table), DSLAdapter.fetchJsonRecords(table)).subscribe(recs -> {
             ctx.verify(() -> Assertions.assertEquals(7, recs.size()));
             cp.flag();
         }, ctx::failNow);
