+++ integtest/pg/src/test/java/io/github/zero88/integtest/jooqx/pg/jooq/PgLegacyRelationTest.java
@@ -4,7 +4,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.jooq.InsertResultStep;
-import org.jooq.exception.SQLStateClass;
 import org.jooq.impl.DSL;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
@@ -17,18 +16,16 @@ import io.github.zero88.jooqx.DSLAdapter;
 import io.github.zero88.jooqx.JsonRecord;
 import io.github.zero88.jooqx.spi.jdbc.JDBCErrorConverterProvider;
 import io.github.zero88.jooqx.spi.pg.PgSQLLegacyTest;
-import io.github.zero88.sample.model.pgsql.tables.Authors;
+import io.github.zero88.sample.model.pgsql.routines.Add;
 import io.github.zero88.sample.model.pgsql.tables.Books;
 import io.github.zero88.sample.model.pgsql.tables.records.AuthorsRecord;
 import io.github.zero88.sample.model.pgsql.tables.records.BooksRecord;
-import io.vertx.core.AsyncResult;
-import io.vertx.core.Handler;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxTestContext;
 
-class PgLeGRelationTest extends PgSQLLegacyTest implements PgUseJooqType, JDBCErrorConverterProvider {
+class PgLegacyRelationTest extends PgSQLLegacyTest implements PgUseJooqType, JDBCErrorConverterProvider {
 
     @Override
     @BeforeEach
@@ -38,9 +35,13 @@ class PgLeGRelationTest extends PgSQLLegacyTest implements PgUseJooqType, JDBCEr
     }
 
     @Test
+    void should_unsupported_routine() {
+        Assertions.assertThrows(UnsupportedOperationException.class, () -> jooqx.routine(new Add()));
+    }
+
+    @Test
     void test_query(VertxTestContext ctx) {
-        final Books table = schema().BOOKS;
-        jooqx.fetchMany(dsl -> dsl.selectFrom(table), ar -> assertResultSize(ctx, ar, 7));
+        jooqx.fetchMany(dsl -> dsl.selectFrom(schema().BOOKS), ar -> assertResultSize(ctx, ar, 7));
     }
 
     @Test
@@ -84,54 +85,7 @@ class PgLeGRelationTest extends PgSQLLegacyTest implements PgUseJooqType, JDBCEr
     }
 
     @Test
-    void test_transaction_multiple_update_but_one_failed(VertxTestContext ctx) {
-        final Checkpoint flag = ctx.checkpoint();
-        final Books table = schema().BOOKS;
-        final Handler<AsyncResult<BooksRecord>> asserter = ar -> {
-            assertJooqException(ctx, ar, SQLStateClass.C23_INTEGRITY_CONSTRAINT_VIOLATION);
-            jooqx.execute(dsl -> dsl.selectFrom(table).where(table.ID.eq(1)), DSLAdapter.fetchOne(table),
-                          ar2 -> ctx.verify(() -> {
-                              final BooksRecord record = assertSuccess(ctx, ar2);
-                              Assertions.assertEquals("The Catcher in the Rye", record.getTitle());
-                              flag.flag();
-                          }));
-        };
-        jooqx.transaction()
-             .run(tx -> tx.execute(dsl -> dsl.update(table)
-                                             .set(DSL.row(table.TITLE), DSL.row("something"))
-                                             .where(table.ID.eq(1))
-                                             .returning(), DSLAdapter.fetchOne(table))
-                          .flatMap(r -> tx.execute(dsl -> dsl.update(table)
-                                                             .set(DSL.row(table.TITLE), DSL.row((String) null))
-                                                             .where(table.ID.eq(2))
-                                                             .returning(), DSLAdapter.fetchOne(table))), asserter);
-    }
-
-    @Test
-    void test_transaction_batch_insert_failed(VertxTestContext ctx) {
-        final Checkpoint flag = ctx.checkpoint();
-        final Authors table = schema().AUTHORS;
-        AuthorsRecord i1 = new AuthorsRecord().setName("n1").setCountry("AT");
-        AuthorsRecord i2 = new AuthorsRecord().setName("n2");
-        final BindBatchValues bindValues = new BindBatchValues().register(table.NAME, table.COUNTRY).add(i1, i2);
-        jooqx.transaction()
-             .run(tx -> tx.batch(tx.dsl().insertInto(table).set(bindValues.getDummyValues()), bindValues), result -> {
-                 // Inconsistent msg in pg < 14 and pg >=14
-                 //                 "Batch entry 1 insert into \"public\".\"authors\" (\"name\", " +
-                 //                 "\"country\") values ('n2', NULL) was aborted: ERROR: null value" +
-                 //                 " in column \"country\" violates not-null constraint\n" +
-                 //                 "  Detail: Failing row contains (10, n2, null).  Call " +
-                 //                 "getNextException to see other errors in the batch."
-                 assertJooqException(ctx, result, SQLStateClass.C23_INTEGRITY_CONSTRAINT_VIOLATION);
-                 jooqx.execute(jooqx.dsl().selectFrom(table), DSLAdapter.fetchMany(table), ar2 -> {
-                     assertResultSize(ctx, ar2, 8);
-                     flag.flag();
-                 });
-             });
-    }
-
-    @Test
-    void test_select_block(VertxTestContext ctx) {
+    void test_block_select(VertxTestContext ctx) {
         final Checkpoint flag = ctx.checkpoint();
         jooqx.block(dsl -> BlockQuery.create()
                                      .add(dsl.selectFrom(schema().AUTHORS).limit(3), DSLAdapter.fetchMany(schema().AUTHORS))
