+++ integtest/pg/src/test/java/io/github/zero88/integtest/jooqx/pg/jooq/PgPoolComplexQueryTest.java
@@ -18,15 +18,13 @@ import io.github.zero88.sample.model.pgsql.Public;
 import io.github.zero88.sample.model.pgsql.tables.pojos.Authors;
 import io.github.zero88.sample.model.pgsql.tables.pojos.Books;
 import io.github.zero88.sample.model.pgsql.tables.records.AuthorsRecord;
-import io.github.zero88.sample.model.pgsql.tables.records.BooksAuthorsRecord;
-import io.github.zero88.sample.model.pgsql.tables.records.BooksRecord;
 import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.Checkpoint;
 import io.vertx.junit5.VertxTestContext;
 import io.vertx.pgclient.PgPool;
 
-class PgReAComplexQueryTest extends PgSQLJooqxTest<PgPool> implements PgPoolProvider, PgUseJooqType {
+class PgPoolComplexQueryTest extends PgSQLJooqxTest<PgPool> implements PgPoolProvider, PgUseJooqType {
 
     @Override
     @BeforeEach
@@ -113,50 +111,6 @@ class PgReAComplexQueryTest extends PgSQLJooqxTest<PgPool> implements PgPoolProv
             Assertions.assertEquals(book2.getTitle(), "The Catcher in the Rye");
             flag.flag();
         }));
-    }
-
-    @Test
-    void test_transaction_insert_into_3_tables(VertxTestContext ctx) {
-        final Checkpoint flag = ctx.checkpoint();
-        final DSLContext dsl = jooqx.dsl();
-        final Public schema = schema().PUBLIC;
-        AuthorsRecord a1 = new AuthorsRecord().setName("Lukas").setCountry("Ger");
-        BooksRecord b1 = new BooksRecord().setTitle("jOOQ doc");
-        jooqx.transaction()
-             .run(tx -> tx.execute(dsl.insertInto(schema.AUTHORS, schema.AUTHORS.NAME, schema.AUTHORS.COUNTRY)
-                                      .values(a1.value2(), a1.value3())
-                                      .returning(), DSLAdapter.fetchOne(schema.AUTHORS))
-                          .flatMap(r1 -> tx.execute(
-                                               dsl.insertInto(schema.BOOKS, schema.BOOKS.TITLE).values(b1.value2()).returning(),
-                                               DSLAdapter.fetchOne(schema.BOOKS))
-                                           .flatMap(r2 -> tx.execute(
-                                               dsl.insertInto(schema.BOOKS_AUTHORS, schema.BOOKS_AUTHORS.BOOK_ID,
-                                                              schema.BOOKS_AUTHORS.AUTHOR_ID)
-                                                  .values(r2.getId(), r1.getId())
-                                                  .returning(), DSLAdapter.fetchOne(schema.BOOKS_AUTHORS)))))
-             .onComplete(ar -> ctx.verify(() -> {
-                 final BooksAuthorsRecord record = assertSuccess(ctx, ar);
-                 Assertions.assertNotNull(record);
-                 Assertions.assertEquals(12, record.getId());
-                 Assertions.assertEquals(9, record.getAuthorId());
-                 Assertions.assertEquals(8, record.getBookId());
-                 final SelectConditionStep<Record> query = dsl.select(schema.AUTHORS.asterisk(),
-                                                                      schema.BOOKS.ID.as("book_id"),
-                                                                      schema.BOOKS.TITLE.as("book_title"))
-                                                              .from(schema.AUTHORS)
-                                                              .join(schema.BOOKS_AUTHORS)
-                                                              .on(schema.AUTHORS.ID.eq(schema.BOOKS_AUTHORS.AUTHOR_ID))
-                                                              .join(schema.BOOKS)
-                                                              .on(schema.BOOKS.ID.eq(schema.BOOKS_AUTHORS.BOOK_ID))
-                                                              .where(schema.BOOKS_AUTHORS.ID.eq(ar.result().getId()));
-                 jooqx.execute(query, DSLAdapter.fetchJsonRecord(query.asTable()), ar2 -> ctx.verify(() -> {
-                     final JsonRecord<?> jsonRecord = assertSuccess(ctx, ar2);
-                     Assertions.assertEquals(new JsonObject("{\"id\":9,\"name\":\"Lukas\",\"country\":\"Ger\"," +
-                                                            "\"book_id\":8,\"book_title\":\"jOOQ doc\"}"),
-                                             jsonRecord.toJson());
-                     flag.flag();
-                 }));
-             }));
     }
 
 }
