--- integtest/pg/src/test/java/io/github/zero88/integtest/jooqx/pg/jooq/PgReAGeometricTest.java
@@ -1,52 +1,0 @@
-package io.github.zero88.integtest.jooqx.pg.jooq;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import io.github.zero88.integtest.jooqx.pg.PgUseJooqType;
-import io.github.zero88.jooqx.DSLAdapter;
-import io.github.zero88.jooqx.spi.pg.PgPoolProvider;
-import io.github.zero88.jooqx.spi.pg.PgSQLErrorConverterProvider;
-import io.github.zero88.jooqx.spi.pg.PgSQLJooqxTest;
-import io.github.zero88.sample.model.pgsql.tables.AllDataTypes;
-import io.github.zero88.sample.model.pgsql.tables.records.AllDataTypesRecord;
-import io.vertx.core.Vertx;
-import io.vertx.junit5.Checkpoint;
-import io.vertx.junit5.VertxTestContext;
-import io.vertx.pgclient.PgPool;
-
-class PgReAGeometricTest extends PgSQLJooqxTest<PgPool>
-    implements PgSQLErrorConverterProvider, PgPoolProvider, PgUseJooqType {
-
-    @Override
-    @BeforeEach
-    public void tearUp(Vertx vertx, VertxTestContext ctx) {
-        super.tearUp(vertx, ctx);
-        this.prepareDatabase(ctx, this, connOpt, "pg_data/geometric.sql");
-    }
-
-    @Test
-    void queryGeometric(VertxTestContext ctx) {
-        final Checkpoint cp = ctx.checkpoint();
-        final AllDataTypes table = schema().ALL_DATA_TYPES;
-        jooqx.execute(dsl -> dsl.selectFrom(table).where(table.ID.eq(71)).limit(1), DSLAdapter.fetchOne(table),
-                      ar -> ctx.verify(() -> {
-                          final AllDataTypesRecord record = assertSuccess(ctx, ar);
-                          System.out.println(record);
-
-                          Assertions.assertNotNull(record.getFPoint());
-
-                          Assertions.assertNotNull(record.getFPolygon());
-                          Assertions.assertNotNull(record.getFBox());
-
-                          Assertions.assertNotNull(record.getFCircle());
-                          Assertions.assertNotNull(record.getFLine());
-                          Assertions.assertNotNull(record.getFLseg());
-                          Assertions.assertNotNull(record.getFClosedPath());
-                          Assertions.assertNotNull(record.getFOpenedPath());
-                          cp.flag();
-                      }));
-    }
-
-}
