--- integtest/pg/src/test/java/io/github/zero88/integtest/jooqx/pg/jooq/PgReAUDTTest.java
@@ -1,81 +1,0 @@
-package io.github.zero88.integtest.jooqx.pg.jooq;
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import io.github.zero88.integtest.jooqx.pg.EnumMoodConverter;
-import io.github.zero88.integtest.jooqx.pg.EnumWeatherConverter;
-import io.github.zero88.integtest.jooqx.pg.FullAddressConverter;
-import io.github.zero88.integtest.jooqx.pg.PgUseJooqType;
-import io.github.zero88.jooqx.DSLAdapter;
-import io.github.zero88.jooqx.datatype.DataTypeMapperRegistry;
-import io.github.zero88.jooqx.datatype.UserTypeAsJooqType;
-import io.github.zero88.jooqx.spi.pg.PgPoolProvider;
-import io.github.zero88.jooqx.spi.pg.PgSQLErrorConverterProvider;
-import io.github.zero88.jooqx.spi.pg.PgSQLJooqxTest;
-import io.github.zero88.sample.model.pgsql.enums.Mood;
-import io.github.zero88.sample.model.pgsql.enums.Weather;
-import io.github.zero88.sample.model.pgsql.tables.AllDataTypes;
-import io.github.zero88.sample.model.pgsql.tables.records.AllDataTypesRecord;
-import io.github.zero88.sample.model.pgsql.udt.records.FullAddressRecord;
-import io.vertx.core.Vertx;
-import io.vertx.junit5.Checkpoint;
-import io.vertx.junit5.VertxTestContext;
-import io.vertx.pgclient.PgPool;
-
-class PgReAUDTTest extends PgSQLJooqxTest<PgPool>
-    implements PgSQLErrorConverterProvider, PgPoolProvider, PgUseJooqType {
-
-    @Override
-    @BeforeEach
-    public void tearUp(Vertx vertx, VertxTestContext ctx) {
-        super.tearUp(vertx, ctx);
-        this.prepareDatabase(ctx, this, connOpt, "pg_data/udt.sql");
-    }
-
-    @Override
-    public DataTypeMapperRegistry typeMapperRegistry() {
-        return PgUseJooqType.super.typeMapperRegistry()
-                                  .addByColumn(schema().ALL_DATA_TYPES.F_UDT_MOOD,
-                                               UserTypeAsJooqType.create(new EnumMoodConverter()))
-                                  .addByColumn(schema().ALL_DATA_TYPES.F_UDT_WEATHER,
-                                               UserTypeAsJooqType.create(new EnumWeatherConverter()))
-                                  .addByColumn(schema().ALL_DATA_TYPES.F_UDT_ADDRESS,
-                                               UserTypeAsJooqType.create(new FullAddressConverter()));
-    }
-
-    @Test
-    void queryEnum(VertxTestContext ctx) {
-        Checkpoint flag = ctx.checkpoint();
-        final AllDataTypes table = schema().ALL_DATA_TYPES;
-        jooqx.execute(dsl -> dsl.selectFrom(table).where(table.ID.eq(43)).limit(1), DSLAdapter.fetchOne(table),
-                      ar -> ctx.verify(() -> {
-                          final AllDataTypesRecord record = assertSuccess(ctx, ar);
-                          System.out.println(record);
-
-                          Assertions.assertEquals(Mood.ok, record.getFUdtMood());
-                          Assertions.assertEquals(Weather.sunny, record.getFUdtWeather());
-                          flag.flag();
-                      }));
-    }
-
-    @Test
-    void queryCustom(VertxTestContext ctx) {
-        Checkpoint flag = ctx.checkpoint();
-        final AllDataTypes table = schema().ALL_DATA_TYPES;
-        jooqx.execute(dsl -> dsl.selectFrom(table).where(table.ID.eq(41)).limit(1), DSLAdapter.fetchOne(table),
-                      ar -> ctx.verify(() -> {
-                          final AllDataTypesRecord record = assertSuccess(ctx, ar);
-                          final FullAddressRecord address = record.getFUdtAddress();
-                          Assertions.assertNotNull(address);
-                          Assertions.assertEquals("US Open", address.getState());
-                          Assertions.assertEquals("Any,town", address.getCity());
-                          Assertions.assertEquals("", address.getStreet());
-                          Assertions.assertEquals(10, address.getNoa());
-                          Assertions.assertEquals(true, address.getHome());
-                          flag.flag();
-                      }));
-    }
-
-}
